tweet_sentiment/requirements.txt
Flask
Flask-Cors
Flask-RESTful
flask-restplus
pandas
onnxruntime
Cython
nemo_toolkit['all']